Create MET - Italian
MET - Italian is a great introductory profile that will be close to a traditional espresso machine. A fast fill that transitions to a constant flow rate will quickly build pressure while maintaining puck resistance. Adjust the End Flow rate to achieve a contact time and extraction that works for you. Adjust grind size to achieve a 6-7 bar peak pressure.